requirements.txt
# Core automation libraries
pyautogui>=0.9.54
pynput>=1.7.6
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR libraries
pytesseract>=0.3.10
easyocr>=1.7.0

# AI API client
openai>=1.3.0
anthropic>=0.8.0

# Configuration and utilities
PyYAML>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0